+++ packages/client/src/index.spec.ts
@@ -7,10 +7,10 @@ describe(`Given the ${LensClient.name} configured to work with staging environme
   };
   const lensClient = new LensClient(config);
 
-  describe(`when ${LensClient.prototype.generateChallenge.name} is called`, () => {
+  describe(`when a method from Auth module is called`, () => {
     it(`returns the challenge for an address`, async () => {
       const address = '0x3fC47cdDcFd59dce20694b575AFc1D94186775b0';
-      const result = await lensClient.generateChallenge(address);
+      const result = await lensClient.auth.generateChallenge(address);
 
       expect(result).toContain('Sign in with ethereum to lens');
     });
